+++ src/components/InkEditor/components/AiAutocomplete/index.ts
@@ -1,0 +1,8 @@
+export { AiAutocomplete } from './AiAutocomplete';
+export type { AiAutocompleteProps, AiAutocompletePosition } from './AiAutocomplete.types';
+export {
+  AI_AUTOCOMPLETE_DEBOUNCE_MS,
+  AI_AUTOCOMPLETE_MAX_PREFIX_CHARS,
+  AI_AUTOCOMPLETE_COMMAND_INSERT_TEXT,
+} from './AiAutocomplete.const';
+export { getTextBeforeCaret, getCaretPositionInRoot } from './AiAutocomplete.utils';