+++ src/constants/index.ts
@@ -11,6 +11,7 @@ export * from './findReplaceMenu.const';
 export * from './blocks.const';
 export * from './defaults.const';
 export * from './aiModels.const';
+export * from './openai.const';
 export * from './premium.const';
 export * from './defaultIcons';
 export * from './contextMenu.const';