+++ src/constants/keyboard.const.ts
@@ -1,3 +1,4 @@
 export const KEY_ENTER = 'Enter';
 export const KEY_ESCAPE = 'Escape';
+export const KEY_TAB = 'Tab';
 export const KEY_Z = 'z';